healthKit 和 后台运行

+++ Function/Function/Tools/KeyBoard/MyTextView.m
@@ -131,7 +131,7 @@
     }else
     {
         face = [NSEntityDescription insertNewObjectForEntityForName:NSStringFromClass([RecentFace class]) inManagedObjectContext:self.context];
-        face.faceIndex = [NSString stringWithFormat:@"%ld",index];
+        face.faceIndex = [NSString stringWithFormat:@"%ld",(long)index];
         face.date = [NSDate date];
     }
